Add option for whether to draw shadow

+++ src/ribbon.ts
@@ -55,6 +55,7 @@ export function generateRibbons(data: { [k: string]: FormDataEntryValue; }) : st
     const ribbonHeight = parseFloat(data["ribbon-height"] as string) / 100 * imageSize
     const threadMargin = parseFloat(data["thread-margin"] as string) / 100 * ribbonHeight
     const shadowHeight = parseFloat(data["shadow-height"] as string) / 100 * ribbonHeight
+    const drawShadow = data["draw-shadow"]
     for (let i = 0; i < 4; ++i) {
         const diagram = new SVGDiagram()
         diagram.setFixedViewBox(0, 0, imageSize, imageSize)
@@ -69,14 +70,16 @@ export function generateRibbons(data: { [k: string]: FormDataEntryValue; }) : st
         } else {
             centerY += ribbonShift;
         }
-        const shadow = diagram.addNode("shadow")
-        shadow.setShape(SVGNode.SHAPE_RECT)
-        shadow.setColor("none")
-        shadow.setCenter(imageSize / 2, centerY)
-        shadow.setFixedSize(1000, shadowHeight)
-        shadow.setGradientAngle(90)
-        shadow.setFillColor("#00000000:#000000A8:#00000000")
-        shadow.delete()
+        if (drawShadow == "on") {
+            const shadow = diagram.addNode("shadow")
+            shadow.setShape(SVGNode.SHAPE_RECT)
+            shadow.setColor("none")
+            shadow.setCenter(imageSize / 2, centerY)
+            shadow.setFixedSize(1000, shadowHeight)
+            shadow.setGradientAngle(90)
+            shadow.setFillColor("#00000000:#000000A8:#00000000")
+            shadow.delete()
+        }
         const ribbon = diagram.addNode("ribbon")
         ribbon.setShape(SVGNode.SHAPE_RECT)
         ribbon.setCenter(imageSize / 2, centerY)
